django_sirius/requirements.txt
djangorestframework==3.15.1
Django==5.0.6
python-dotenv==1.0.1
django-extensions==3.2.3